+++ src/user_interface/include/user_interface.hh
@@ -76,8 +76,8 @@ private:
     struct CurrentTrip
     {
         uint32_t start_distance {0};
-        uint32_t start_wh_consumed {0};
-        uint32_t start_wh_regenerated {0};
+        float start_wh_consumed {0};
+        float start_wh_regenerated {0};
     };
 
     void OnStartup() final;
